+++ web/documentserver-example/nodejs/helpers/docManager.js
@@ -381,6 +381,10 @@ DocManager.prototype.getInternalExtension = function getInternalExtension(fileTy
     return '.pptx';
   }
 
+  if (fileType === fileUtility.fileType.pdf) { // .pptx for pdf type
+    return '.pdf';
+  }
+
   return '.docx'; // the default value is .docx
 };