+++ web/documentserver-example/nodejs/helpers/fileUtility.js
@@ -64,6 +64,7 @@ fileUtility.fileType = {
   word: 'word',
   cell: 'cell',
   slide: 'slide',
+  pdf: 'pdf',
 };
 
 fileUtility.getSuppotredExtensions = function getSuppotredExtensions() {